Move locationsFor to awards-for-all
